requirements.txt
streamlit==1.28.1
Pillow==10.0.1
requests==2.31.0
google-generativeai==0.3.2
